test resource stats and nlp tasks working *

+++ package-admin/src/pt_pump_up_admin/CRUD.py
@@ -4,57 +4,64 @@ from requests import Request
 
 class CRUD(ABC):
 
-    def __init__(self, route, identifier: int = None) -> None:
+    def __init__(self, route, *args, **kwargs) -> None:
         if route is None:
             raise ValueError("route cannot be None")
 
         self.route = route.replace("/", "")
-        self.identifier = identifier
-        self.json = None
+        self.json = dict()
+
+        for arg in args:
+            self.json.update(arg)
+
+        for key, value in kwargs.items():
+            self.json[key] = value
 
     def index(self) -> Request:
-        return Request(
+        return self, Request(
             method="GET",
             url=self.route,
         )
 
     def store(self) -> Request:
-        return Request(
+        return self, Request(
             method="POST",
             url=self.route,
+            json=self.json
         )
 
-    def show(self, identifier: int = None) -> Request:
-
+    def show(self) -> Request:
+        """
         if self.identifier is None and identifier is None:
             raise ValueError("identifier cannot be None")
         elif self.identifier is None and identifier is not None:
             self.identifier = identifier
-        return Request(
+        """
+
+        return self, Request(
             method="GET",
-            url=f"{self.route}/{self.identifier}",
+            url=f"{self.route}/{self.json.id}",
         )
 
-    def update(self, identifier: int = None) -> Request:
-
+    def update(self) -> Request:
+        """
         if self.identifier is None and identifier is None:
             raise ValueError("identifier cannot be None")
         elif self.identifier is None and identifier is not None:
             self.identifier = identifier
+        """
 
-        return Request(
+        return self, Request(
             method="PUT",
-            url=f"{self.route}/{self.identifier}",
+            url=f"{self.route}/{self.json.id}",
         )
 
-    def delete(self, identifier: int = None) -> Request:
+    def destroy(self) -> Request:
 
-        if self.identifier is None and identifier is None:
-            raise ValueError("identifier cannot be None")
-        elif self.identifier is None and identifier is not None:
-            self.identifier = identifier
+        if self.json['id'] is None:
+            raise ValueError("id cannot be None")
 
-        return Request(
+        return self, Request(
             method="DELETE",
-            url=f"{self.route}/{self.identifier}",
+            url=f"{self.route}/{self.json['id']}",
         )
